+++ js/const.mjs
@@ -1,0 +1,3 @@
+const CNT_PHOTOS = 25;
+
+export {CNT_PHOTOS};
